Refactor trading MVP architecture by consolidating main script functionality into core modules. Removed legacy main.py and backend main.py files, enhancing separation of concerns with dedicated CLI, configuration, and orchestration components. Improved logging and error handling throughout the application.
